--- gardener/provider_types.go
@@ -1,6 +1,0 @@
-package gardener
-
-const (
-	aws string = "aws"
-	gcp string = "gcp"
-)
