+++ main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"github.com/hashicorp/terraform/plugin"
 	"github.com/hashicorp/terraform/terraform"
-	"github.com/kyma-incubator/terraform-provider-gardener/gardener"
+	"github.com/kyma-incubator/terraform-provider-gardener/provider"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() terraform.ResourceProvider {
-			return gardener.Provider()
+			return provider.New()
 		},
 	})
 }
