+++ shoot/gcp.go
@@ -1,10 +1,16 @@
-package gardener
+package shoot
 
 import (
+	"strconv"
+
+	gardencorev1alpha1 "github.com/gardener/gardener/pkg/apis/core/v1alpha1"
+	gardener_types "github.com/gardener/gardener/pkg/apis/garden/v1beta1"
 	"github.com/hashicorp/terraform/helper/schema"
 )
 
-func resourceGCPShoot() *schema.Resource {
+const gcp string = "gcp"
+
+func GCPShoot() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceGCPServerCreate,
 		Read:   resourceServerRead,
@@ -79,4 +85,30 @@ func resourceGCPServerCreate(d *schema.ResourceData, m interface{}) error {
 
 func resourceGCPServerUpdate(d *schema.ResourceData, m interface{}) error {
 	return resourceServerUpdate(d, m, gcp)
+}
+
+func createGCPSpec(spec gardener_types.ShootSpec, d *schema.ResourceData, secretBinding string) gardener_types.ShootSpec {
+	spec.Cloud.SecretBindingRef.Name = secretBinding
+	spec.Cloud.GCP = &gardener_types.GCPCloud{
+		Networks: gardener_types.GCPNetworks{
+			Workers: []gardencorev1alpha1.CIDR{"10.250.0.0/19"}, // TODO replace hardcoded
+		},
+		Workers: getGCPWorkers(d),
+		Zones:   getZones(d),
+	}
+	return spec
+}
+
+func getGCPWorkers(d *schema.ResourceData) []gardener_types.GCPWorker {
+	numWorkers := d.Get("worker.#").(int)
+	resultWorkers := make([]gardener_types.GCPWorker, numWorkers)
+	for i := 0; i < numWorkers; i++ {
+		var worker = "worker." + strconv.Itoa(i)
+		resultWorkers[i] = gardener_types.GCPWorker{
+			Worker:     createGardenWorker(worker, d),
+			VolumeSize: d.Get(worker + ".volumesize").(string),
+			VolumeType: d.Get(worker + ".volumetype").(string),
+		}
+	}
+	return resultWorkers
 }
